@@ -1,5 +1,11 @@
-module github.com/lieturd/parse-template
+module github.com/cocreators-ee/parse-template
 
-go 1.13
+go 1.24.1
 
 require github.com/stretchr/testify v1.4.0
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
+)
